Renamed app in sub-projects to lm_agent and lm_backend
