+++ agent/lm_agent/api/booking_proxy.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends
 
-from app.forward import ForwardOperation
+from lm_agent.forward import ForwardOperation
 
 router = APIRouter()
