+++ agent/lm_agent/tokenstat.py
@@ -9,11 +9,14 @@ from shlex import quote
 
 from pydantic import BaseModel, Field
 
-from app.backend_utils import get_config_from_backend
-from app.config import ENCODING, PRODUCT_FEATURE_RX, TOOL_TIMEOUT, settings
-from app.logs import logger
-from app.parsing import flexlm
-from app.workload_managers.slurm.cmd_utils import get_tokens_for_license, sacctmgr_modify_resource
+from lm_agent.backend_utils import get_config_from_backend
+from lm_agent.config import ENCODING, PRODUCT_FEATURE_RX, TOOL_TIMEOUT, settings
+from lm_agent.logs import logger
+from lm_agent.parsing import flexlm
+from lm_agent.workload_managers.slurm.cmd_utils import (
+    get_tokens_for_license,
+    sacctmgr_modify_resource,
+)
 
 
 class LicenseService(BaseModel):
