+++ agent/pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 description = "Provids an agent for interacting with license manager"
 authors = ["OmniVector Solutions <info@omnivector.solutions>"]
 license = "MIT"
-packages = [{ include = "app" }]
+packages = [{ include = "lm_agent" }]
 
 [tool.poetry.dependencies]
 python = "^3.8"
@@ -36,7 +36,7 @@ types-click = "^7.1.5"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-addopts = "--random-order --cov=app --cov-report=term-missing --cov-fail-under=70"
+addopts = "--random-order --cov=lm_agent --cov-report=term-missing --cov-fail-under=70"
 testpaths = ["tests"]
 
 [tool.black]
@@ -44,15 +44,15 @@ line-length = 100
 
 [tool.isort]
 line_length = 100
-src_paths = ["app", "tests"]
+src_paths = ["lm_agent", "tests"]
 multi_line_output = 3
 include_trailing_comma = true
 
 [tool.poetry.scripts]
 lm-configure = "bin.lmconfigure:app"
 lm-create-jwt = "bin.createjwt:main"
-slurmctld-prolog = "app.workload_managers.slurm.slurmctld_prolog:main"
-slurmctld-epilog = "app.workload_managers.slurm.slurmctld_epilog:main"
+slurmctld-prolog = "lm_agent.workload_managers.slurm.slurmctld_prolog:main"
+slurmctld-epilog = "lm_agent.workload_managers.slurm.slurmctld_epilog:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
