+++ agent/tests/workload_managers/slurm/test_cmd_utils.py
@@ -1,6 +1,9 @@
 from pytest import raises
 
-from app.workload_managers.slurm.cmd_utils import SqueueParserUnexpectedInputError, squeue_parser
+from lm_agent.workload_managers.slurm.cmd_utils import (
+    SqueueParserUnexpectedInputError,
+    squeue_parser,
+)
 
 
 def test_squeue_with_bad_input():
