+++ backend/alembic/env.py
@@ -3,7 +3,7 @@ from logging.config import fileConfig
 from alembic import context
 from sqlalchemy import engine_from_config, pool
 
-from app.table_schemas import metadata as lm_backend_metadata
+from lm_backend.table_schemas import metadata as lm_backend_metadata
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
