+++ backend/tests/api/test_booking.py
@@ -1,10 +1,10 @@
 from httpx import AsyncClient
 from pytest import fixture, mark
 
-from app import table_schemas
-from app.api import booking
-from app.api_schemas import ConfigurationRow
-from app.storage import database
+from lm_backend import table_schemas
+from lm_backend.api import booking
+from lm_backend.api_schemas import ConfigurationRow
+from lm_backend.storage import database
 
 
 @fixture
