+++ backend/tests/test_debug.py
@@ -3,10 +3,10 @@ from unittest.mock import patch
 from fastapi.exceptions import HTTPException
 from pytest import raises
 
-from app import debug
+from lm_backend import debug
 
 
-def test_debug():
+def test_debug(backend_client):
     """
     Does the debug dependency raise an exception when the app is *not* in debug mode?
     """
